[runtime-security] extract self tester to own file
