+++ pkg/security/module/module.go
@@ -591,36 +591,3 @@ func NewModule(cfg *sconfig.Config) (module.Module, error) {
 
 	return m, nil
 }
-
-// SelfTester represents all the state needed to conduct rule injection test at startup
-type SelfTester struct {
-	Enabled         bool
-	waitingForEvent bool
-	EventChan       chan eval.Event
-}
-
-// NewSelfTester returns a new SelfTester, enabled or not
-func NewSelfTester(enabled bool) *SelfTester {
-	return &SelfTester{
-		Enabled:         enabled,
-		waitingForEvent: false,
-		EventChan:       make(chan eval.Event),
-	}
-}
-
-// BeginWaitingForEvent passes the tester in the waiting for event state
-func (t *SelfTester) BeginWaitingForEvent() {
-	t.waitingForEvent = true
-}
-
-// EndWaitingForEvent exits the waiting for event state
-func (t *SelfTester) EndWaitingForEvent() {
-	t.waitingForEvent = false
-}
-
-// SendEvent sends an event to the tester
-func (t *SelfTester) SendEventIfExpecting(event eval.Event) {
-	if t.Enabled && t.waitingForEvent {
-		t.EventChan <- event
-	}
-}
